requirements.txt
Django==3.0.3
mysqlclient==1.4.6
python-dotenv==0.11.0
django-crispy-forms==1.8.1
asgiref==3.5.2
certifi==2022.9.14
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
django-crontab==0.7.1
django-import-export==2.0.2
django-model-utils==4.0.0
django-mysql==3.4.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.21.3
et-xmlfile==1.1.0
idna==2.10
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pyparsing==3.0.9
python3-saml==1.9.0
pytz==2022.2.1
PyYAML==6.0
requests==2.24.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.25.11
xlrd==2.0.1
xlwt==1.3.0
xmlsec==1.3.13